Too much.
Refactored the preprocess.py file to use more techniques
and wrapped it all in a class.

Changed the config file and added extra parameters.

Added sota.py which has a few state of the art
architectures that I import into the train.py
also made the model name defined in the config file
change what model is being trained.

Changes to be committed:
	modified:   .gitignore
	modified:   configs/mi_config.yaml
	modified:   configs/ssvep_config.yaml
	new file:   notebooks/yet_another_messy_notebook.ipynb
	deleted:    run/MI/EEGNet-20250625_024659-inprogress/checkpoints/eegnet-mi-best-f1-val_f1=0.4419-epoch=00.ckpt
	deleted:    run/MI/EEGNet-20250625_024659-inprogress/checkpoints/eegnet-mi-best-f1-val_f1=0.4419-epoch=03.ckpt
	deleted:    run/MI/EEGNet-20250625_024659-inprogress/checkpoints/eegnet-mi-best-f1-val_f1=0.4419-epoch=09.ckpt
	deleted:    run/MI/EEGNet-20250625_024659-inprogress/checkpoints/last.ckpt
	modified:   scripts/preprocess_data.py
	modified:   src/bci_aic3/config.py
	modified:   src/bci_aic3/models/eegnet.py
	new file:   src/bci_aic3/models/sota.py
	modified:   src/bci_aic3/preprocess.py
	modified:   src/bci_aic3/train.py
	modified:   training_stats/mi_stats.pt
	modified:   training_stats/ssvep_stats.pt